Merge d'avant vacs : exception handler

+++ filRouge/src/main/java/com/m2i/filRouge/service/AbstractGenericService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 
 import com.m2i.filRouge.converter.GenericConverter;
+import com.m2i.filRouge.exception.NotFoundException;
 
 public abstract class AbstractGenericService<E,ID,DTO> implements GenericService<E,ID,DTO>{
 
@@ -23,17 +24,26 @@ public abstract class AbstractGenericService<E,ID,DTO> implements GenericService
 	}
 	
 	// pour pas utiliser le converter dans notre rest controller et direct faire un findDtoByid dans la methode Get
-	public DTO findDtoById(ID id) {
-		return GenericConverter.map(this.findById(id), getDtoClass());
+	public DTO findDtoById(ID id) throws NotFoundException {
+		E e = this.findById(id);
+		if(e != null) {
+			return GenericConverter.map(e,getDtoClass());
+		} else {
+			throw new NotFoundException("Entity not found for id "+id);
+		}
+		
 	}
 
 	@Override
 	public E save(E entity) {
 		return dao.save(entity);
 	}
-
+	
 	@Override
-	public void deleteById(ID id) {
+	public void deleteById(ID id) throws NotFoundException{
+		
+		if(!(dao.existsById(id)))
+			throw new NotFoundException("no entity to delete for id: "+id);
 		dao.deleteById(id);
 	}
